MID-30 separate drowing from game engine
