+++ packages/client/src/game/containers/apple-worm-game/apple-worm-game.scss
@@ -5,14 +5,6 @@
   gap: 10px 20px;
 }
 
-.aw-game-board {
-  position: relative;
-  background-color: #fff;
-  border: 1px dashed rgb(51 51 51 / 20%);
-  margin: 20px auto;
-  background-color: transparent;
-}
-
 .aw-game-overlay {
   position: fixed;
   top: 50%;
