+++ packages/client/src/game/ui/game-canvas/game-canvas.scss
@@ -1,0 +1,7 @@
+.aw-game-canvas {
+  position: relative;
+  background-color: #fff;
+  border: 1px dashed rgb(51 51 51 / 20%);
+  margin: 20px auto;
+  background-color: transparent;
+}
