+++ packages/client/src/pages/game-page/game-page.scss
