+++ packages/client/src/pages/game-page/index.ts
@@ -1,1 +1,1 @@
-export * from './game-page/game-page';
+export * from './game-page';
