+++ packages/client/src/widgets/game-header/game-board.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export function GameHeader() {
   return <div>GameHeader Component</div>;
 }
